non_blocking network io interface for rusty hermit
This introduces an entire new network interface for rusty hermit
it tries to resemble epoll and and berkley sockets.

A socket is created by sys_socket
is bound to a purpose by sys_{tcp,event,waker}_bind
and can then be used by passing it to corresponding functions.

Every socket has a non_blocking option that, if applied,
causes io actions to return error of kind io::ErrorKind::WouldBlock
instead of blocking the thread.

To use a non_blocking api effectively (no busy waiting or unconditional polling)
the event socket type can be used to combine many sockets into a single
object which can be blocked on until IO occures.

The waker socket is a socket with events that are not based in io
but in active calls from the user.
This allows for users to manually wake a blocking call to
sys_event_wait.

Sharing types between RustyHermit and HermitAbi led to hermit-sys
depending on hermit-abi.